create example workload with stateful set
[#150790628]

+++ docker-images/persistent-example/app.rb
@@ -6,10 +6,14 @@ set :bind, '0.0.0.0'
 get '/' do
     log_file = ENV['LOG_FILE']
 
-    File.write(log_file, "")
-
-    last_line = IO.readlines(log_file)[-1]
-    last_number = last_line.nil? ? 0 : last_line.strip.to_i
+    if File.exist? log_file
+        last_line = IO.readlines(log_file)[-1]
+        last_number = last_line.strip.to_i
+    else
+        puts "File doesn't exist, creating..."
+        File.write(log_file, "")        
+        last_number = 0
+    end
 
     num_lines = File.open(log_file, "r").readlines.size
